requirements.txt

pillow
keras
tensorflow
numpy
opencv-contrib-python
